fast refresh and auth handling selector split
